Added button show leaderboard & update player score.

+++ app/src/main/java/org/fransis/game/words/Player.java
@@ -1,13 +1,13 @@
 package org.fransis.game.words;
 
 public class Player {
-    private int score = 0;
+    private long score = 0;
 
-    public int getScore() {
+    public long getScore() {
         return score;
     }
 
-    public int addScore(int delta){
+    public long addScore(long delta){
         if(delta > 0) {
             this.score += delta;
         }
